% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{del.stopwords}
\alias{del.stopwords}
\title{Removes stop words for a string the language of which is known}
\usage{
del.stopwords(str, lang)
}
\arguments{
\item{str}{The string which to delete the stop words from}

\item{lang}{Either an ISO 639-2/3 or a language name from which to derive a ISO 639-2 code. For language names performs string matching.}
}
\value{
A character vector corresponding to the string \code{str} without stopwords for the language \code{lang}
}
\description{
Removes stop words for a string the language of which is known
}
\keyword{internal}
