% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopwords.R
\name{languages}
\alias{languages}
\title{Returns all the ISO 639-1 code available in the Stopwords ISO library}
\usage{
languages()
}
\value{
A data frame with a row for each supported languages and columns for the several ISO code (639-2, 639-3, 639-1) and the name.
}
\description{
See the relevant \href{https://en.wikipedia.org/wiki/ISO_639-1}{Wikipedia article} for details of the language code.
}
