% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{match.lang}
\alias{match.lang}
\title{Matches a string with the ISO 639-1 code available in this library}
\usage{
match.lang(lang)
}
\arguments{
\item{lang}{Either an ISO 639-2/3 or a language name from which to derive a ISO 639-2 code. For language names performs string matching.}
}
\value{
A character vector containing the two-letter ISO 639-1 code associated to the requested language.
}
\description{
See \url{https://en.wikipedia.org/wiki/ISO_639-1} for details of the language code.
}
\keyword{internal}
